refactor(user): remove participat page & update sidebar
